requirements.txt
connexion == 2.2.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
python-jose[cryptography]
Flask-RESTful==0.3.6
flask-testing
connexion[swagger-ui]
requests==2.19.1
Flask-Cors==3.0.7
werkzeug==0.15.5
